+++ .github/workflows/update-ranking.yml
@@ -26,11 +26,6 @@ jobs:
         with:
           python-version: "3.10"
 
-      - name: Install dependencies
-        run: |
-          pip install pipenv
-          pipenv install --dev
-
       - name: Run the ranking update script
         run: pipenv run python -m ranking_calculator
