+++ ranking_calculator/player.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from ranking_calculator.config import MAX_COUNTING_TOURNAMENTS, TOURNAMENT_BASE_POINTS
+from ranking_calculator.config import MAX_COUNTING_TOURNAMENTS
 from ranking_calculator.tournament import Tournament
 
 
@@ -11,12 +11,7 @@ class TournamentPlacements:
 
     @property
     def points(self) -> float:
-        base_points = TOURNAMENT_BASE_POINTS.get(self.rank, 0)
-        return (
-            base_points
-            * self.tournament.player_multiplier
-            * self.tournament.age_multiplier
-        )
+        return self.tournament.get_points(self.rank)
 
 
 @dataclass
@@ -37,3 +32,22 @@ class Player:
 @dataclass
 class RankedPlayer(Player):
     rank: int
+
+
+class PlayerList:
+    def __init__(self) -> None:
+        self.players: list[Player] = []
+
+    def get_or_create_player(self, player_name: str) -> Player:
+        player = next((p for p in self.players if p.name == player_name), None)
+        if player is None:
+            player = Player(name=player_name, tournament_placements=[])
+            self.players.append(player)
+        return player
+
+    def update_playerlist_from_tournament(self, tournament: Tournament):
+        for result in tournament.tournament_results:
+            player = self.get_or_create_player(result.player_name)
+            player.tournament_placements.append(
+                TournamentPlacements(tournament=tournament, rank=result.rank)
+            )
