+++ ranking_calculator/ranking_system.py
@@ -2,20 +2,22 @@ import copy
 from datetime import datetime
 
 from ranking_calculator.config import PLAYER_MULTIPLIERS
-from ranking_calculator.player import Player, RankedPlayer, TournamentPlacements
+from ranking_calculator.player import PlayerList, RankedPlayer
 from ranking_calculator.tournament import Tournament
 
 
 class RankingSystem:
-    def __init__(self):
-        self.players: list[Player] = []
+    def __init__(self) -> None:
+        self.player_list: PlayerList = PlayerList()
         self.tournament_history: list[Tournament] = []
         self.ranking_history: dict[datetime, list[RankedPlayer]] = {}
 
     @property
     def ranked_players(self) -> list[RankedPlayer]:
         sorted_players = sorted(
-            self.players, key=lambda player: player.ranking_points, reverse=True
+            self.player_list.players,
+            key=lambda player: player.ranking_points,
+            reverse=True,
         )
         return [
             RankedPlayer(
@@ -39,29 +41,15 @@ class RankingSystem:
                 multiplier += PLAYER_MULTIPLIERS.get(player_rank, 0.0)
         return multiplier
 
-    def get_or_create_player(self, player_name: str) -> Player:
-        player = next((p for p in self.players if p.name == player_name), None)
-        if player is None:
-            player = Player(name=player_name, tournament_placements=[])
-            self.players.append(player)
-        return player
-
-    def add_player_tournament_placements(self, tournament: Tournament):
-        for result in tournament.tournament_results:
-            player = self.get_or_create_player(result.player_name)
-            player.tournament_placements.append(
-                TournamentPlacements(tournament=tournament, rank=result.rank)
-            )
-
-    def add_tournament(self, tournament: Tournament):
-        # First we update the ranking with the new age multipliers
+    def update_ranking_with_tournament(self, tournament: Tournament):
+        # First we update the tournament's age multiplier
         self.update_age_multipliers(tournament.date)
         # Then we calculate the player multiplier for the tournament
         tournament.player_multiplier = self.calculate_tournament_player_multiplier(
             tournament
         )
         # Then we update the player's placements with the tournament results
-        self.add_player_tournament_placements(tournament)
-        self.tournament_history.append(tournament)
+        self.player_list.update_playerlist_from_tournament(tournament)
         # Finally, we store the ranking history
+        self.tournament_history.append(tournament)
         self.ranking_history[tournament.date] = copy.deepcopy(self.ranked_players)
