--- ranking_calculator/rules.json
@@ -1,81 +1,0 @@
-{
-    "max_counting_tournaments": 3,
-    "player_multipliers": {
-        "10": 0.1,
-        "20": 0.06,
-        "30": 0.03,
-        "50": 0.005
-    },    
-    "age_multipliers": {
-        "12": 1.0,
-        "18": 0.8,
-        "24": 0.6,
-        "48": 0.4
-    },
-    "blank_points": {
-        "1": 700,
-        "2": 630,
-        "3": 560,
-        "4": 490,
-        "5": 420,
-        "6": 420,
-        "7": 350,
-        "8": 350,
-        "9": 280,
-        "10": 280,
-        "11": 280,
-        "12": 280,
-        "13": 210,
-        "14": 210,
-        "15": 210,
-        "16": 210,
-        "17": 140,
-        "18": 140,
-        "19": 140,
-        "20": 140,
-        "21": 140,
-        "22": 140,
-        "23": 140,
-        "24": 140,
-        "25": 105,
-        "26": 105,
-        "27": 105,
-        "28": 105,
-        "29": 105,
-        "30": 105,
-        "31": 105,
-        "32": 105,
-        "33": 70,
-        "34": 70,
-        "35": 70,
-        "36": 70,
-        "37": 70,
-        "38": 70,
-        "39": 70,
-        "40": 70,
-        "41": 70,
-        "42": 70,
-        "43": 70,
-        "44": 70,
-        "45": 70,
-        "46": 70,
-        "47": 70,
-        "48": 70,
-        "49": 70,
-        "50": 70,
-        "51": 70,
-        "52": 70,
-        "53": 70,
-        "54": 70,
-        "55": 70,
-        "56": 70,
-        "57": 70,
-        "58": 70,
-        "59": 70,
-        "60": 70,
-        "61": 70,
-        "62": 70,
-        "63": 70,
-        "64": 70
-    }
-}
