--- tests/test_data/1900-01-01_Sydney.csv
@@ -1,7 +1,0 @@
-name,rank,category
-Albert Einstein,1,advanced
-Henk Henkerson,1,advanced
-Piet Pieterson,2,advanced
-Marie Curie,2,advanced
-Isaac Newton,1,intermediate
-Galileo Galilei,1,intermediate
